+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.3.0 — 2014-xx-xx
+- New command `xtc doctor` checks and if needed attempts to fix the project setup. Currently fixes broken or missing symlink to generator-xtc.
+- Re-implemented and expanded generator-xtc symlinking code in xtc-cli.
 - New `command `xtc ls`. Lists versions published to npm.
 
 ## 0.2.2 — 2014-06-13